- начало создания класса и объектов Item
