+++ src/main/java/ru/practicum/shareit/item/model/Item.java
@@ -1,7 +1,27 @@
 package ru.practicum.shareit.item.model;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
 /**
- * TODO Sprint add-controllers.
+ * В пакете item создайте класс Item и добавьте в него следующие поля:
+ * id — уникальный идентификатор вещи;
+ * name — краткое название;
+ * description — развёрнутое описание;
+ * available — статус о том, доступна или нет вещь для аренды;
+ * owner — владелец вещи;
+ * request — если вещь была создана по запросу другого пользователя, то в этом поле будет храниться ссылка на соответствующий запрос.
  */
+
+@Data
+@Builder
+@AllArgsConstructor
 public class Item {
+    private Long id;
+    private String name;
+    private String description;
+    private Boolean available;
+    private Long owner;
+    private Long request;
 }
