+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.item.service;
+
+public interface ItemService {
+}
