+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.item.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ItemServiceImpl implements ItemService {
+}
