+++ src/main/java/ru/practicum/shareit/item/storage/ItemRepository.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.item.storage;
+
+public interface ItemRepository {
+}
