+++ src/main/java/ru/practicum/shareit/item/storage/inMemory/ItemRepositoryImpl.java
@@ -1,0 +1,8 @@
+package ru.practicum.shareit.item.storage.inMemory;
+
+import org.springframework.stereotype.Repository;
+import ru.practicum.shareit.item.storage.ItemRepository;
+
+@Repository
+public class ItemRepositoryImpl implements ItemRepository {
+}
