+++ src/main/java/ru/practicum/shareit/user/model/User.java
@@ -1,0 +1,16 @@
+package ru.practicum.shareit.user.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+@AllArgsConstructor
+public class User {
+    private Long id;
+    private String name;
+    private String email;
+}
+
+
