+++ src/main/java/ru/practicum/shareit/user/storge/UserRepository.java
@@ -1,4 +1,6 @@
-package ru.practicum.shareit.user;
+package ru.practicum.shareit.user.storge;
+
+import ru.practicum.shareit.user.dto.UserDto;
 
 import java.util.List;
