+++ src/main/java/ru/practicum/shareit/user/utils/UserMapper.java
@@ -1,7 +1,9 @@
-package ru.practicum.shareit.user;
+package ru.practicum.shareit.user.utils;
 
 import lombok.Builder;
 import lombok.Data;
+import ru.practicum.shareit.user.dto.UserDto;
+import ru.practicum.shareit.user.model.User;
 
 @Data
 @Builder
